+++ common/src/main/java/fr/loudo/narrativecraft/client/editors/cameraangle/CameraAngleMakerEditorCameraRenderer.java
@@ -45,6 +45,9 @@ public class CameraAngleMakerEditorCameraRenderer {
         EditorMaker editorMaker =
                 ClientNarrativeCraftMod.getInstance().getPlayerSession().getEditor();
         if (!(editorMaker instanceof ClientCameraAngleMakerEditorMaker cameraAngleEditor)) return;
+
+        cameraAngleEditor.renderAnchorPoint(poseStack);
+
         if (cameraAngleEditor.getPreviewCamera() != null) return;
 
         Minecraft minecraft = Minecraft.getInstance();