+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 - Added 'space' type for text tag
+- Change the anchor point of a dialog for a camera view
 
 ### Fixes
-
 - Dialog resizing crackling animation first second